pyproject.toml
[project]
name = "ai-dev3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncio>=3.4.3",
    "fastembed>=0.4.2",
    "httpx>=0.27.2",
    "langfuse>=2.53.9",
    "markdownify>=0.13.1",
    "neo4j>=5.26.0",
    "openai>=1.53.1",
    "parsel>=1.9.1",
    "pydantic>=2.9.2",
    "python-decouple>=3.8",
    "qdrant-client>=1.12.1",
    "requests>=2.32.3",
    "termcolor>=2.5.0",
    "uuid>=1.30",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
polygon = "ai_dev3.polygon:main"
S01E01 = "ai_dev3.S01E01.anti_captcha:main"
S01E02 = "ai_dev3.S01E02.veryfi:main"
S01E03 = "ai_dev3.S01E03.json_report:main"
S01E05 = "ai_dev3.S01E05.censorship:task"
S02E01 = "ai_dev3.S02E01.mp3:main"
S02E02 = "ai_dev3.S02E02.cities:task"
S02E03 = "ai_dev3.S02E03.robotid:main"
S02E04 = "ai_dev3.S02E04.categories:main"
S02E05 = "ai_dev3.S02E05.main:main"
S03E01 = "ai_dev3.S03E01.main:main"
S03E02 = "ai_dev3.S03E02.main:main"
S03E03 = "ai_dev3.S03E03.main:main"
S03E04 = "ai_dev3.S03E04.main:main"
S03E05 = "ai_dev3.S03E05.main:main"
S04E01 = "ai_dev3.S04E01.main:main"
S04E02 = "ai_dev3.S04E02.main:main"
S04E03 = "ai_dev3.S04E03.main:main"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
